Refactored ParseUtility.java to ParserUtil.java
